added sudoku class.
